7.0.x

+++ modules/apps/web-experience/export-import/export-import-service/src/main/java/com/liferay/exportimport/controller/LayoutImportController.java
@@ -344,10 +344,8 @@ public class LayoutImportController implements ImportController {
 			(Map<Long, Layout>)portletDataContext.getNewPrimaryKeysMap(
 				Layout.class + ".layout");
 
-		if (_log.isDebugEnabled()) {
-			if (!portletElements.isEmpty()) {
-				_log.debug("Deleting portlet data");
-			}
+		if (_log.isDebugEnabled() && !portletElements.isEmpty()) {
+			_log.debug("Deleting portlet data");
 		}
 
 		for (Element portletElement : portletElements) {
@@ -696,7 +694,11 @@ public class LayoutImportController implements ImportController {
 					sourcePrototypeLayoutUuid, layoutSetPrototype.getGroupId(),
 					true);
 
-				if (sourcePrototypeLayout == null) {
+				if ((sourcePrototypeLayout == null) &&
+					(_layoutLocalService.fetchLayout(
+						layout.getUuid(), layout.getGroupId(),
+						layout.isPrivateLayout()) != null)) {
+
 					_layoutLocalService.deleteLayout(
 						layout, false, serviceContext);
 				}
@@ -708,10 +710,8 @@ public class LayoutImportController implements ImportController {
 
 		List<Element> layoutElements = layoutsElement.elements();
 
-		if (_log.isDebugEnabled()) {
-			if (!layoutElements.isEmpty()) {
-				_log.debug("Importing layouts");
-			}
+		if (_log.isDebugEnabled() && !layoutElements.isEmpty()) {
+			_log.debug("Importing layouts");
 		}
 
 		List<String> sourceLayoutsUuids = new ArrayList<>();
@@ -722,10 +722,8 @@ public class LayoutImportController implements ImportController {
 
 		// Import portlets
 
-		if (_log.isDebugEnabled()) {
-			if (!portletElements.isEmpty()) {
-				_log.debug("Importing portlets");
-			}
+		if (_log.isDebugEnabled() && !portletElements.isEmpty()) {
+			_log.debug("Importing portlets");
 		}
 
 		Map<Long, Layout> layouts =
@@ -1206,7 +1204,7 @@ public class LayoutImportController implements ImportController {
 					}
 					catch (NoSuchLayoutException nsle) {
 						if (_log.isWarnEnabled()) {
-							_log.warn(nsle);
+							_log.warn(nsle, nsle);
 						}
 					}
 				}
